Main file
Correction of conventional structural drawings using AutoLISP, a case study.

This program uses geometrical properties of a conventional building drawings to achieve its goals.
Then, it Corrects the placement of steel rebars in reinforced-concrete columns in the first floor.
Due to conventional available length of steel rebars used in structures, sometimes using two separate sets of steel bars in columns of two subsequent storeys is not possible. Hence, the drawing has to be manually corrected. In simple cases this is not a problem. When more corrections need to be made, however, using an automated approach is essential to preserve time and costs. This programs shows, in a simple case, how to transform two separate steel rebars into one, in columns of the first and second floors in a conventional reinforced-concrete building.